# src/main/gen/src/main/java/io/security/portfolio/api/RolesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.security.portfolio.api;

import io.security.portfolio.dto.CreateRole201Response;
import io.security.portfolio.dto.ListRoles200Response;
import io.security.portfolio.dto.RoleRequest;
import io.security.portfolio.dto.UpdateUserRolesRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-07T12:21:57.339446635+09:00[Asia/Seoul]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "Roles", description = "the Roles API")
public interface RolesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/roles : Role 생성
     *
     * @param roleRequest  (required)
     * @return 생성된 Role (status code 201)
     */
    @Operation(
        operationId = "createRole",
        summary = "Role 생성",
        tags = { "Roles" },
        responses = {
            @ApiResponse(responseCode = "201", description = "생성된 Role", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateRole201Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/roles",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<CreateRole201Response> createRole(
        @Parameter(name = "RoleRequest", description = "", required = true) @Valid @RequestBody RoleRequest roleRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/users/{userId}/roles : 특정 유저의 글로벌 Role 조회
     *
     * @param userId  (required)
     * @return 유저에 부여된 Role 목록 (status code 200)
     */
    @Operation(
        operationId = "getUserRoles",
        summary = "특정 유저의 글로벌 Role 조회",
        tags = { "Roles" },
        responses = {
            @ApiResponse(responseCode = "200", description = "유저에 부여된 Role 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListRoles200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/users/{userId}/roles",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ListRoles200Response> getUserRoles(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") Long userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 }, { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 } ], \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/roles : 전체 Role 목록 조회
     *
     * @return Role 목록 (status code 200)
     */
    @Operation(
        operationId = "listRoles",
        summary = "전체 Role 목록 조회",
        tags = { "Roles" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Role 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListRoles200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/roles",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ListRoles200Response> listRoles(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 }, { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 } ], \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /api/users/{userId}/roles : 특정 유저의 글로벌 Role 변경
     *
     * @param userId  (required)
     * @param updateUserRolesRequest  (required)
     * @return 변경된 Role 목록 (status code 200)
     */
    @Operation(
        operationId = "updateUserRoles",
        summary = "특정 유저의 글로벌 Role 변경",
        tags = { "Roles" },
        responses = {
            @ApiResponse(responseCode = "200", description = "변경된 Role 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListRoles200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/users/{userId}/roles",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ListRoles200Response> updateUserRoles(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") Long userId,
        @Parameter(name = "UpdateUserRolesRequest", description = "", required = true) @Valid @RequestBody UpdateUserRolesRequest updateUserRolesRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 }, { \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0 } ], \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}