# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Document Authorization API
  version: 1.0.0
  description: |
    RBAC + Team Role + ABAC + ACL(PBAC 기반) 구조를 가진 문서 인가 데모 API.
    - 글로벌 RBAC: users / roles / user_roles
    - 팀/조직: teams / team_members
    - 문서: documents
    - ACL(문서별 권한): document_permissions
    - 감사 로그: audit_logs
    - 인증: refresh_tokens
servers:
  - url: http://localhost:8080
    description: Local dev server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        message:
          type: string
          nullable: true
          description: 추가 메시지 (없으면 null)
        data:
          nullable: true
          description: 실제 응답 데이터 (엔드포인트별로 타입 상이)

    # =====================
    # Auth
    # =====================
...
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: "예: SYSTEM_ADMIN, ORG_ADMIN, USER"
        description:
          type: string
          nullable: true
    RoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    UpdateUserRolesRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds:
          type: array
          items:
            type: integer
            format: int64
          description: "해당 유저에게 부여할 글로벌 Role ID 목록"

    # =====================
    # Team / TeamMember
    # =====================
...
paths:
...
  /api/roles:
    get:
      tags: [Roles]
      summary: 전체 Role 목록 조회
      operationId: listRoles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Role 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RoleResponse'
    post:
      tags: [Roles]
      summary: Role 생성
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '201':
          description: 생성된 Role
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RoleResponse'
  /api/users/{userId}/roles:
    get:
      tags: [Roles]
      summary: 특정 유저의 글로벌 Role 조회
      operationId: getUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 유저에 부여된 Role 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RoleResponse'
    put:
      tags: [Roles]
      summary: 특정 유저의 글로벌 Role 변경
      operationId: updateUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRolesRequest'
      responses:
        '200':
          description: 변경된 Role 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RoleResponse'

  # =====================
  # Teams / TeamMembers
  # =====================
...